+++ crates/dies-core/src/world.rs
@@ -12,7 +12,7 @@ pub enum GameState {
     Timeout,
     Stop,
     PrepareKickoff,
-    BallReplacement(Vector2<f32>),
+    BallReplacement(Vector2<f64>),
     PreparePenalty,
     Kickoff,
     FreeKick,
@@ -28,9 +28,9 @@ pub struct BallData {
     /// seconds). This is the time that ssl-vision received the frame.
     pub timestamp: f64,
     /// Position of the ball filtered by us, in mm, in dies coordinates
-    pub position: Vector3<f32>,
+    pub position: Vector3<f64>,
     /// Velocity of the ball in mm/s, in dies coordinates
-    pub velocity: Vector3<f32>,
+    pub velocity: Vector3<f64>,
 }
 
 #[derive(Serialize, Clone, Debug, Default)]
@@ -51,16 +51,16 @@ pub struct PlayerData {
     /// The player's unique id
     pub id: u32,
     /// Unfiltered position as reported by vision
-    pub raw_position: Vector2<f32>,
+    pub raw_position: Vector2<f64>,
     /// Position of the player filtered by us in mm, in dies coordinates
-    pub position: Vector2<f32>,
+    pub position: Vector2<f64>,
     /// Velocity of the player in mm/s, in dies coordinates
-    pub velocity: Vector2<f32>,
+    pub velocity: Vector2<f64>,
     /// Orientation of the player, in radians, (`-pi`, `pi`), where `0` is the positive
     /// x direction, and `pi/2` is the positive y direction.
-    pub orientation: f32,
+    pub orientation: f64,
     /// Angular speed of the player (in rad/s)
-    pub angular_speed: f32,
+    pub angular_speed: f64,
 }
 
 /// A struct to store the world state from a single frame.
